refactor: split pavilion.ts into modular components
- Extract types into types.ts for better type management
- Split transaction builders into transaction-builders.ts
- Move kiosk utilities to kiosk-utils.ts
- Separate scene config functions to scene-config.ts
- Maintain backward compatibility via re-exports
